requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.8.1
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.3.2
Django==5.0
django-phonenumber-field==8.0.0
django-twilio==0.14.3.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
frozenlist==1.4.1
idna==3.8
multidict==6.0.5
phonenumbers==8.13.45
psycopg2-binary==2.9.9
PyJWT==2.9.0
requests==2.32.3
sqlparse==0.5.1
twilio==9.3.0
urllib3==2.2.2
yarl==1.11.0
